feat: rework debounce directive and add signal input modifier mechanism
Here is my new approach that should make this more modular.
It's inspired from Angular's ControlValueAccessor.
The idea is that we allow directives that implement the
SignalInputModifier interface and register themselves as
a SIGNAL_INPUT_MODIFIER.

These directives will be notified whenever the
onModelChange on the SignalInputDirective is called and are
can store the SignalInputDirective's this.formField.value.set
locally and update the signal of the input whenever it is
appropriate.

Only one modifier is SIGNAL_INPUT_MODIFIER per input as we
cannot know which oder of execution is desired, etc.