feat: introduce reset function for FormField and FormGroup
For FormFields this will reset the value to the initial value and also
set the state to PRISTINE and UNTOUCHED.
For FormGroups this will go through every control and execute the reset
function of said control. If the control happens to be an array, it will
reset the array to a copy of the initial array.

closes #11