fix: change signal input modifier to accept whole value signal instead of set function
BREAKING CHANGE: we replace the registerOnSet function of the SignalInputModifier with a new
registerValueSignal function. We do this, because just assigning the set function to a new property
within our class can break (and did break) due to it changing the scope from where set is called.
Specifically, we are not calling ourSignal.set anymore this is ourSignal, but instead
SignalInputDebounceDirective.set where this is now SignalInputDebounceDirective. To fix this we simply
store the whole signal containing the value of our input and call the valueSignal.set in the debounce
directive. To enable this our SignalInputModifier interface changes from registerOnSet to registerValueSignal.